refactor: rename date/time folders to date-picker/time-picker and update related imports
