+++ docs/reference-guides/data/data-core-annotations.md
@@ -12,7 +12,8 @@ This package is still experimental. “Experimental” means this is an early im
 
 Nothing to document.
 
-<!-- END TOKEN(Autogenerated selectors|../../../packages/annotations/src/store/selectors.ts) -->
+
+<!-- END TOKEN(Autogenerated selectors|../../../packages/annotations/src/store/selectors.js) -->
 
 ## Actions
