+++ docs/reference-guides/data/data-core-edit-post.md
@@ -138,13 +138,13 @@ _Returns_
 
 ### isEditingTemplate
 
-> **Deprecated**
+> **Deprecated** 
 
 Returns true if the template editing mode is enabled.
 
 ### isEditorPanelEnabled
 
-> **Deprecated**
+> **Deprecated** 
 
 Returns true if the given panel is enabled, or false otherwise. Panels are enabled by default.
 
@@ -159,7 +159,7 @@ _Returns_
 
 ### isEditorPanelOpened
 
-> **Deprecated**
+> **Deprecated** 
 
 Returns true if the given panel is open, or false otherwise. Panels are closed by default.
 
@@ -174,7 +174,7 @@ _Returns_
 
 ### isEditorPanelRemoved
 
-> **Deprecated**
+> **Deprecated** 
 
 Returns true if the given panel was programmatically removed, or false otherwise. All panels are not removed by default.
 
@@ -214,7 +214,7 @@ _Returns_
 
 ### isInserterOpened
 
-> **Deprecated**
+> **Deprecated** 
 
 Returns true if the inserter is opened.
 
@@ -306,7 +306,7 @@ _Returns_
 
 ### isPublishSidebarOpened
 
-> **Deprecated**
+> **Deprecated** 
 
 Returns true if the publish sidebar is opened.
 
@@ -329,6 +329,7 @@ _Parameters_
 _Returns_
 
 -   `boolean`: Whether the metaboxes are being saved.
+
 
 <!-- END TOKEN(Autogenerated selectors|../../../packages/edit-post/src/store/selectors.js) -->
