+++ packages/components/src/date-time/index.ts
@@ -1,8 +1,8 @@
 /**
  * Internal dependencies
  */
-import { default as DatePicker } from './date';
-import { default as TimePicker } from './time';
+import { default as DatePicker } from './date-picker';
+import { default as TimePicker } from './time-picker';
 import { default as DateTimePicker } from './date-time';
 
 export { DatePicker, TimePicker };
